Added Wall Powers
testing initial functionality of wall powers. Players will be able to
use them during battles like weapons that are part of the wall itself.
- Also changed unit HP to float to manage it better as a %

+++ TheWall_0.0/Assets/Scripts/Units/AI_Enemy.cs
@@ -21,7 +21,7 @@ public class AI_Enemy : Battle_Unit {
 		description = "Default monster thing.";
 		// Random stat init, ** This is for now
 		myStats = initStats (quality);
-		hitPoints = myStats [0];
+		hitPoints = (float)myStats [0];
 		attackRating = myStats [1];
 		defenseRating = myStats [2];
