+++ TheWall_0.0/Assets/Scripts/Units/Captain.cs
@@ -24,7 +24,7 @@ public class Captain : Battle_Unit {
 		description = "Default captain guy";
 		// Random stat init, ** This is for now
 		myStats = initStats (quality);
-		hitPoints = myStats [0];
+		hitPoints = (float)myStats [0];
 		attackRating = myStats [1];
 		defenseRating = myStats [2];
